cloud/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
google-genai>=1.0.0
